refactor send button
