+++ packages/onchainkit/src/wallet/components/wallet-advanced-send/components/SendButton.tsx
@@ -4,7 +4,6 @@ import type { PortfolioTokenWithFiatValue } from '@/api/types';
 import type { LifecycleStatusUpdate } from '@/internal/types';
 import { Transaction } from '@/transaction/components/Transaction';
 import { TransactionButton } from '@/transaction/components/TransactionButton';
-import { useTransactionContext } from '@/transaction/components/TransactionProvider';
 import { TransactionStatus } from '@/transaction/components/TransactionStatus';
 import { TransactionStatusAction } from '@/transaction/components/TransactionStatusAction';
 import { TransactionStatusLabel } from '@/transaction/components/TransactionStatusLabel';
@@ -14,14 +13,12 @@ import type {
 } from '@/transaction/types';
 import { useCallback } from 'react';
 import { parseUnits } from 'viem';
-import { type Chain, base } from 'viem/chains';
+import { base } from 'viem/chains';
 import { useWalletContext } from '../../WalletProvider';
 import type { SendLifecycleStatus } from '../types';
-import { defaultSendTxSuccessHandler } from '../utils/defaultSendTxSuccessHandler';
 import { getSendCalldata } from '../utils/getSendCalldata';
 import { useSendContext } from './SendProvider';
-import { cn, pressable, text } from '@/styles/theme';
-import { Spinner } from '@/internal/components/Spinner';
+import { RenderSendButton } from './RenderSendButton';
 
 export function SendButton() {
   const { chain: senderChain, isSponsored } = useWalletContext();
@@ -76,9 +73,11 @@ export function SendButton() {
       calls={calldata ? [calldata] : []}
       onStatus={handleStatus}
     >
-      <SendTransactionButton
-        label={buttonLabel}
-        senderChain={senderChain}
+      <TransactionButton
+        text={buttonLabel}
+        render={(params: TransactionButtonRenderParams) => (
+          <RenderSendButton {...params} label={buttonLabel} />
+        )}
         disabled={disableSendButton}
       />
       <TransactionStatus>
@@ -86,93 +85,6 @@ export function SendButton() {
         <TransactionStatusAction />
       </TransactionStatus>
     </Transaction>
-  );
-}
-
-/**
- * SendTransactionButton required to be a nested component in order to pull from TransactionContext.
- * Need to pull from TransactionContext in order to get transactionHash and transactionId.
- * Need transactionHash and transactionId in order to determine where to open the transaction in the wallet or explorer.
- */
-function SendTransactionButton({
-  label,
-  senderChain,
-  disabled,
-}: {
-  label: string;
-  senderChain?: Chain | null;
-  disabled?: boolean;
-}) {
-  const { address, setActiveFeature } = useWalletContext();
-  const { transactionHash, transactionId } = useTransactionContext();
-
-  const completionHandler = useCallback(() => {
-    setActiveFeature(null);
-  }, [setActiveFeature]);
-
-  const customRender = useCallback(
-    ({ context, onSubmit }: TransactionButtonRenderParams) => {
-      const classNames = cn(
-        pressable.primary,
-        'rounded-ock-default',
-        'w-full rounded-xl',
-        'px-4 py-3 font-medium leading-6',
-        disabled && pressable.disabled,
-        text.headline,
-        'text-ock-text-foreground',
-      );
-      if (context.receipt) {
-        return (
-          <button
-            className={classNames}
-            onClick={() =>
-              defaultSendTxSuccessHandler({
-                transactionId,
-                transactionHash,
-                senderChain: senderChain ?? undefined,
-                address: address ?? undefined,
-                onComplete: completionHandler,
-              })(context.receipt)
-            }
-            disabled={disabled}
-          >
-            View transaction
-          </button>
-        );
-      }
-      if (context.errorMessage) {
-        return (
-          <button className={classNames} onClick={onSubmit} disabled={disabled}>
-            Try again
-          </button>
-        );
-      }
-      if (context.isLoading) {
-        return (
-          <button className={classNames} disabled={disabled}>
-            <Spinner />
-          </button>
-        );
-      }
-      return (
-        <button className={classNames} disabled={disabled} onClick={onSubmit}>
-          {label}
-        </button>
-      );
-    },
-    [
-      address,
-      completionHandler,
-      disabled,
-      label,
-      senderChain,
-      transactionHash,
-      transactionId,
-    ],
-  );
-
-  return (
-    <TransactionButton text={label} render={customRender} disabled={disabled} />
   );
 }
